correction

+++ 0x02-functions_nested_loops/9-times_table.c
@@ -6,11 +6,11 @@
 
 void times_table(void)
 {
-	int i;
+	long i;
 
 	for (i = 0; i <= 9; i++)
 	{
-		int j;
+		long j;
 
 		for (j = 0; j <= 9; j++)
 		{
